refactor vault into own package
