+++ pkg/whisper.go
@@ -16,6 +16,7 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/rotationalio/whisper/pkg/config"
 	"github.com/rotationalio/whisper/pkg/logger"
+	"github.com/rotationalio/whisper/pkg/vault"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 )
@@ -61,7 +62,7 @@ func New(conf config.Config) (s *Server, err error) {
 	// Create the vault to store secrets in (Google Secret Manager)
 	// TODO: if we're in test mode, create a mock secret manager
 	if conf.Mode != gin.TestMode {
-		if s.vault, err = NewSecretManager(conf.Google); err != nil {
+		if s.vault, err = vault.New(conf.Google); err != nil {
 			return nil, err
 		}
 		log.Debug().Msg("connected to google secret manager")
@@ -76,7 +77,7 @@ func New(conf config.Config) (s *Server, err error) {
 	// Add CORS configuration
 	// TODO: configure origins from the environment rather than hard-coding
 	s.router.Use(cors.New(cors.Config{
-        AllowAllOrigins:  true,
+		AllowAllOrigins:  true,
 		AllowMethods:     []string{"GET", "POST", "PUT", "PATCH", "DELETE", "HEAD"},
 		AllowHeaders:     []string{"Origin", "Content-Length", "Content-Type"},
 		AllowCredentials: true,
@@ -103,12 +104,12 @@ func New(conf config.Config) (s *Server, err error) {
 
 type Server struct {
 	sync.RWMutex
-	conf    config.Config  // configuration of the API server
-	srv     *http.Server   // handle to a custom http server with specified API defaults
-	router  *gin.Engine    // the http handler and associated middlware
-	vault   *SecretManager // storage for all secrets the whisper application manages
-	healthy bool           // application state of the server
-	errc    chan error     // synchronize shutdown gracefully
+	conf    config.Config        // configuration of the API server
+	srv     *http.Server         // handle to a custom http server with specified API defaults
+	router  *gin.Engine          // the http handler and associated middlware
+	vault   *vault.SecretManager // storage for all secrets the whisper application manages
+	healthy bool                 // application state of the server
+	errc    chan error           // synchronize shutdown gracefully
 }
 
 func (s *Server) Serve() (err error) {
